+++ rust/dbn-macros/tests/ui/json_serialize_conflicting_dbn_attr.rs
@@ -1,0 +1,10 @@
+use dbn_macros::JsonSerialize;
+
+#[derive(JsonSerialize)]
+#[repr(C)]
+struct Record {
+    #[dbn(fixed_price, unix_nanos)]
+    pub a: u8,
+}
+
+fn main() {}