+++ scripts/build.sh
@@ -1,2 +1,3 @@
 #! /usr/bin/env bash
+cargo --version
 cargo build --all-features